Refactoring to get logging sorted out.
The proc.py was refactored into a bunch of different files.

I had to explicitly provide a file that I explicitly logged to in order
to ensure it was closed when the worker was done. The tradeoff is that
when it cleans up the directory we don't log anything. Technically, it
still logs to stdout, but that is already a noop since we're using a
daemon.

I also refactored out the client to a specific module. I'm trying to use
a functional style here so no classes, avoiding side effects, named
tuples for types and naming modules like I might for classes.